+++ packages/patternfly-3/patternfly-react-extensions/src/components/DiffView/index.js
@@ -1,0 +1,3 @@
+import DiffView from './DiffView';
+
+export { DiffView };