+++ packages/patternfly-4/react-core/src/components/Modal/examples/SimpleModal.js
@@ -14,6 +14,7 @@ class SimpleModal extends React.Component {
 
   render() {
     const { isModalOpen } = this.state;
+    const reactRoot = typeof document !== 'undefined' ? document.querySelector('#___gatsby') : '';
 
     return (
       <React.Fragment>
@@ -32,7 +33,7 @@ class SimpleModal extends React.Component {
               Confirm
             </Button>
           ]}
-          reactRoot={document.querySelector('#___gatsby')}
+          reactRoot={reactRoot}
         >
           Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore
           magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo