+++ packages/patternfly-4/react-core/src/components/index.js
@@ -19,7 +19,8 @@ export * from './List';
 export * from './LoginPage';
 export * from './Modal';
 export * from './Nav';
-export * from './Popover';
+// export * from './Popover'; // Not ready yet
+export * from './Page';
 export * from './Progress';
 export * from './Radio';
 export * from './Select';