+++ packages/patternfly-3/patternfly-react-extensions/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.14.2"></a>
+## [2.14.2](https://github.com/patternfly/patternfly-react/compare/patternfly-react-extensions@2.14.0...patternfly-react-extensions@2.14.2) (2018-12-19)
+
+
+
+
+**Note:** Version bump only for package patternfly-react-extensions
+
 <a name="2.14.0"></a>
 # [2.14.0](https://github.com/patternfly/patternfly-react/compare/patternfly-react-extensions@2.13.19...patternfly-react-extensions@2.14.0) (2018-12-18)