+++ packages/patternfly-4/react-core/CHANGELOG.md
@@ -3,7 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.40.4"></a>
+## [1.40.4](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@1.40.3...@patternfly/react-core@1.40.4) (2018-12-19)
 
+
+### Bug Fixes
+
+* PF4 About Modal: make trademark, logoImageSrc and logoImageAlt p… ([#1082](https://github.com/patternfly/patternfly-react/issues/1082)) ([dbc8b06](https://github.com/patternfly/patternfly-react/commit/dbc8b06))
+
+
+
+
+<a name="1.40.3"></a>
+## [1.40.3](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@1.40.2...@patternfly/react-core@1.40.3) (2018-12-19)
+
+
+### Bug Fixes
+
+* **page:** move page from layouts to components ([#1089](https://github.com/patternfly/patternfly-react/issues/1089)) ([4c3e6af](https://github.com/patternfly/patternfly-react/commit/4c3e6af))
+* update alt text in LoginPage ([#1084](https://github.com/patternfly/patternfly-react/issues/1084)) ([2665351](https://github.com/patternfly/patternfly-react/commit/2665351))
+
+
+
+
+<a name="1.40.2"></a>
+## [1.40.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@1.40.1...@patternfly/react-core@1.40.2) (2018-12-19)
+
+
+
+
+**Note:** Version bump only for package @patternfly/react-core
 
 <a name="1.40.0"></a>
 # [1.40.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@1.39.1...@patternfly/react-core@1.40.0) (2018-12-11)